% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentBowtie2.R
\name{alignmentBowtie2}
\alias{alignmentBowtie2}
\title{Alignment with Bowtie 2}
\usage{
alignmentBowtie2(file, out_name = NULL, suffix_fastq = NULL,
  type = "SE", out_dir = "bam/", cores = 6,
  index = "/biodata/indices/species/Hsapiens/ucsc.hg19",
  path_logs = "logs/", run = TRUE, ...)
}
\arguments{
\item{file}{Character string (single-end) or character vector (paired-end) with the path and
filename of the sample to be analyzed.}

\item{out_name}{Character string indicating the name for the output file. If not provided, will use
suffix_fastq to remove the suffix and use the same name as fastq file.}

\item{suffix_fastq}{Character vector indicating the suffix to remove from fastq file in order to
generate output sam file. Only necessary if out_name is not provided. If this is a paired end sample
it should also include the name of read 1 (example: "_read1.fastq.gz").}

\item{type}{Character string with the type of approach used for the sequencing, either "SE" (default) or "PE".}

\item{out_dir}{Character string indicating the path where the final bam files should be stored.}

\item{cores}{Integer indicating the number of cores to use for the analysis.}

\item{index}{Character string with the path and name for the Bowtie2 index as specified in Bowtie2 manual.}

\item{path_logs}{Character string with the path for the logs to be stored.}

\item{run}{Logical indicating whether to run the alignment (for testing purposes). Default: TRUE}

\item{...}{Further arguments to be passed to Bowtie2 call.}
}
\value{
Writes "out_name.raw.bam" in out_dir. Also generates a log for the alignment (sample.alignment.log) in your path_logs.
}
\description{
Aligns FastQ files to a reference genome using Bowtie2, converts the output to BAM, sorts it
according to genomic position and indexes the final BAM file.
}
\examples{
\dontrun{
alignmentBowtie2(file=paste0(path, "fastq/NL1_h3k27ac_sample.fastq.gz"),
                 suffix_fastq=".fastq.gz",
                 type="SE",
                 index=" /biodata/indices/species/Hsapiens/ucsc.hg19",
                 out_dir=paste0(path, "bam/"),
                 path_logs=paste0(path, "logs/"),
                 cores=6,
                 run=FALSE)
}
}
