% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentTophat.R
\name{alignmentTophat}
\alias{alignmentTophat}
\title{Alignment with TopHat}
\usage{
alignmentTophat(samples, type = "SE", path_fastq, suffix_fastq, path_bam,
  cores, trans_index, index)
}
\arguments{
\item{samples}{Single element or character vector with the name of the samples (whitout any specific file format suffix).}

\item{type}{Character string with the type of approach used for the sequencing, either "SE" (default) or "PE".}

\item{path_fastq}{Character string indicating the path which contains the fastq files.}

\item{suffix_fastq}{Character string with the suffix of the input fastq files (ex. "fq.gz", "fastq.gz").}

\item{path_bam}{Character string indicating the path where the final bam files should be stored.}

\item{cores}{Integer indicating the number of cores to use for the analysis.}

\item{trans_index}{Indexed gene annotation (in our case gencode v18) by TopHat.}

\item{index}{Path with the path and name for the Bowtie2 as specified in Bowtie2 manual.}
}
\value{
Returns a BAM file with the algimnent.
}
\description{
Aligns FastQ files to a reference genome using TopHat (RNA-seq data).
}
\examples{
\dontrun{
alignmentTophat(samples=c("sample1", "sample2"), type="SE", index="/biodata/index/ucsc.hg19",
                path_fastq="~/test/samples/raw/", suffix_fastq="fq.gz", path_bam="~/test/samples/bam/",
                cores=6, trans_index="~/data/gencode.v18.annotation.tophat2index")
}
}
