% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateToNearest.R
\name{annotateToNearest}
\alias{annotateToNearest}
\title{Annotate regions to nearest gene}
\usage{
annotateToNearest(regions, genes, promLimit = 2000)
}
\arguments{
\item{regions}{GRanges including the query regions}

\item{genes}{GRanges including data for genes to be annotated to. Should contain strand info.}

\item{promLimit}{Limit for annotating a region as a promoter. Default: 2kb.}
}
\value{
Returns an annotated data.frame including info of the regions and genes, along with distance to the gene and annotation.
}
\description{
Annotates genomic regions to nearest feature in a GRanges object.
}
