% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToVis.R
\name{bamToBigWig}
\alias{bamToBigWig}
\title{Conversion from BAM to visualization formats}
\usage{
bamToBigWig(
  bamfile,
  scaling_factor = 1,
  path_bw,
  chr_sizes,
  bedtools_genomecov = "bedtools genomecov",
  bdg_to_bw = "bedGraphToBigWig",
  uncompressed_bw = FALSE,
  cores = 5,
  perc_mem = "50\%"
)
}
\arguments{
\item{bamfile}{Path and name of the bam file to convert.}

\item{scaling_factor}{Factor to use for scaling.}

\item{path_bw}{Character string indicating the path where you want to save visualization files (it will be created if it does not exist).}

\item{chr_sizes}{Character string indicating the path where the file with chromosome name and sizes can be found.}

\item{bedtools_genomecov}{Character string indicating the path where the program converter from bedGraph to BigWig can be found. Default assumes it is in your PATH.}

\item{bdg_to_bw}{Character string indicating the path where the bedtools genomecov binary can be found. Default assumes it is in your PATH.}

\item{uncompressed_bw}{Logical indicating whether to compress or not the resulting bigWig file.}

\item{cores}{Number of cores to use.}

\item{perc_mem}{Percentage of memory to use.}
}
\value{
Two visualization files: bedgraph (.bdg) and bigwig (.bw) in your path_bw.
}
\description{
Converts post-processed BAM files to visualization formats such as BedGraph (compressed) and BigWig.
}
