% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToVis.R
\name{bamToVis}
\alias{bamToVis}
\title{Conversion from BAM to visualization formats}
\usage{
bamToVis(samples, summaryStats = "", path_bam, path_vis, scales = "",
  gen_sizes = "~/data/hg19.len",
  path_bedGraphToBigWig = "~/tools/bedGraphToBigWig")
}
\arguments{
\item{samples}{Character string or character vector with the name of the samples (whitout any specific file format suffix).}

\item{summaryStats}{Character string with the location of the summary statistics .rda file.}

\item{path_bam}{Character string indicating the path where the final bam files should be stored.}

\item{path_vis}{Character string indicating the path where you want to save visualization files (it will be created if it does not exist).}

\item{scales}{Number or vector with the scales for the samples. Default is using those in summaryStats.}

\item{gen_sizes}{Character string indicating the path where the file with chromosome name and sizes can be found. Default value: "~/data/hg19.len"}

\item{path_bedGraphToBigWig}{Character string indicating the path where the program converter from bedGraph to BigWig can be found. Default value: "~/tools/bedGraphToBigWig".}
}
\value{
Two visualization files: bedgraph (.bdg) and bigwig (.bw) in your path_vis.
}
\description{
Converts post-processed BAM files to visualization formats such as BedGraph (compressed) and BigWig.
}
\examples{
\dontrun{
bamToVis("thyroid_sample", summaryStats="testData/logs/summary.rda", path_bam="testData/bam/",
         path_vis="testData/vis/")
}
}
