% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanCons.R
\name{calculateMeanCons}
\alias{calculateMeanCons}
\title{Calculates mean conservation score in a set of regions}
\usage{
calculateMeanCons(regions, scope = 500, bin = 50,
  phastConsBW = "~/data/phastCons_46_placentalMammals/phastCons46way.placental.bw")
}
\arguments{
\item{regions}{Set of regions as a \code{GRanges} object to use for the analysis. The first mcols column should either
contain a unique identifier for the region or be empty.}

\item{scope}{Number of bp to extend the regions from the peak center, both upstream
and downstream (Length of region = scope*2). Default=500bp.}

\item{bin}{Number of bp for binning the regions. Default=50bp.}

\item{phastConsBW}{Path and name of the bigWig file containing the phastCons conservation
scores (or any other bigWig for calculating mean score).}
}
\value{
A data.frame object containing the bin position (from the peak center) and
the mean conservation score.
}
\description{
Using a set of regions as input, extends them \code{scope} from the center (upstream
and downstream) and bins them in \code{bins} bp. Then obtains the conservation score from
a phastCons bigWig file, and calculates the mean per bin in all regions.
}
