% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCUTTAG.R
\name{.getBDGforSEACR}
\alias{.getBDGforSEACR}
\title{Obtain bedgraph files for SEACR input}
\usage{
.getBDGforSEACR(
  bam_file,
  cores,
  chr_sizes,
  bedtools_bamtobed,
  bedtools_genomecov
)
}
\description{
@inheritParams peakCallingSEACR
}
