% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{Quality Control with FASTQC}
\usage{
fastqc(files, out_dir = "fastqc/", cores = 6)
}
\arguments{
\item{files}{String or character vector with the path and filename of the fastq samples.}

\item{out_dir}{Character string with the name of the folder where you want to save FastQC results.}

\item{cores}{Integer indicating the number of cores to use for the analysis.}
}
\value{
It creates a folder "out_dir" that contains the html and other files that are generated as a result of running FASTQC over your samples.
}
\description{
Runs FASTQC for quality control analysis of your FastQ reads.
}
\examples{
\dontrun{
fastqc(samples="path/to/fastq/file.fastq.gz", out_dir="fastqc/", cores=6)
}
}
