% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtOutBAM.R
\name{filtOutBAM}
\alias{filtOutBAM}
\title{Filter out reads from BAM file}
\usage{
filtOutBAM(
  file,
  path_logs,
  type = "SE",
  remove = c("chrM", "chrUn", "_random", "_hap", "_gl", "EBV"),
  blacklist = "~/data/consensusBlacklist.bed",
  cores = 6
)
}
\arguments{
\item{file}{Filenme and path for the BAM file to be filtered out.}

\item{path_logs}{Character indicating the output directory for the logs.}

\item{type}{Sequence type, one of "SE" (single end) or "PE" (paired end).}

\item{remove}{Character vector with chr that will be filtered out. Any chromosome
name containing matches for these characters will be removed.}

\item{blacklist}{Character indicating the file containing blacklist regions in
bed format. Any reads overlapping these regions will be discarded.}

\item{cores}{Number of cores to use for the analysis.}
}
\value{
File without the ".raw" preffix containing the final reads, along with its index (.bai).
Also returns invisibly the name of the output file.
}
\description{
Function that filters out the following reads: 1) aligned no non-cannonical chromosomes, 2) not aligned, 3) aligned to ENCODE blacklisted regions and 3) duplicated reads.
}
\examples{
\dontrun{
file <- filtOutBAM(file=file,
                   path_logs=path_logs,
                   cores=8)
}
}
