% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNameFromPath.R
\name{getNameFromPath}
\alias{getNameFromPath}
\title{Get name from path}
\usage{
getNameFromPath(path, suffix = NULL, prefix = NULL)
}
\arguments{
\item{path}{Character string or vector with the path.}

\item{suffix}{Characer string with the suffix you want to remove from the filename. For example: ".bam"}

\item{prefix}{Character string with the prefix you want to remove from the filename. For example: "ATAC_"}
}
\value{
A character string or vector with the filename.
}
\description{
Obtains a file name from a path.
}
\examples{
\dontrun{
getNameFromPath("test/data/cov_file.txt", suffix=".txt", preffix="cov_")
}
}
