% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{Get alignment and processing stats from BAM files}
\usage{
getStats(
  path_logs = "Logs/",
  path_peaks = "Peaks/",
  peak_suffix = "narrowPeak"
)
}
\arguments{
\item{path_logs}{Path where the log files from samtools mrkdup are stored.}

\item{path_peaks}{Path where the peaks are saved.}

\item{peak_suffix}{Type of the peak files, either narrowPeak or broadPeak}
}
\value{
Data.frame with some alignment and processing statistics for all the input samples.
}
\description{
Given a Logs/ folder generated by the processing functions in this package,
returns the statistics for alignment, chrM, duplication rate and number of
peaks called.
}
\examples{
\dontrun{
stats <- getStats(raw_bam=c("path/to/file.bam", "path/to/file2.bam"), path_logs="path/logs")
}
}
