% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTseqCount.R
\name{HTseqCount}
\alias{HTseqCount}
\title{Assigning counts to transcripts with HTseqCount}
\usage{
HTseqCount(samples, bam_path, transcript)
}
\arguments{
\item{samples}{Single element or character vector with the name of the samples (whitout any specific file format suffix).}

\item{transcript}{Transcriptome of reference to use for assigning the counts.}

\item{path_bam}{Character string indicating the path where the final bam files should be stored.}
}
\value{
A file ".htseq.stats" with the general statistics and a file ".htseq.counts" with the number of counts assigned to each transcript.
}
\description{
Using an input BAM file, it assigns the counts to each transcript annotated in the transcriptome of interest.
}
\examples{
\dontrun{
HTseqCount(samples=c("sample1", "sample2"), path_bam="~/test/samples/bam/",
                transcript="~/data/gencode.v18.annotation.gtf")
}
}
