% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBAMs.R
\name{mergeBAMs}
\alias{mergeBAMs}
\title{Merge BAM files}
\usage{
mergeBAMs(files, out_name, path_bam, path_out, cores = 6)
}
\arguments{
\item{files}{Name for the bam files you want to merge.}

\item{out_name}{Character string with the name for the merged BAM file.}

\item{path_bam}{Character string indicating the path where the input bam files are.}

\item{path_out}{Character string indicating the path where the final merged bam files should be stored.}

\item{cores}{Integer specifying the number of cores to use.}
}
\value{
Bed files in path_pc with the location of the enriched regions.
}
\description{
Merges several BAM files in one.
}
\examples{
\dontrun{callPeak()}
}
