% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsetATAC.R
\name{offsetATAC}
\alias{offsetATAC}
\title{Offset correction for ATAC-seq data}
\usage{
offsetATAC(file, cores = 6, gen_sizes = "~/data/hg19.len")
}
\arguments{
\item{file}{Character string with the filename and path for the BAM file.}

\item{cores}{Number of cores to use for the processing.}

\item{gen_sizes}{Character string indicating the path where the file with chromosome name and sizes can be found. Default value: "~/data/hg19.len"}
}
\value{
Returns the final BAM file, sorted and indexed, in path_bam.
}
\description{
It performs the offset correction needed for ATAC-seq data: +4bp in forward strand; -5bp in reverse strand.
}
\examples{
\dontrun{
offsetATAC(file)
}
}
