% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsetATAC.R
\name{offsetATACSE}
\alias{offsetATACSE}
\title{Offset correction for SE ATAC-seq data}
\usage{
offsetATACSE(file, chunk = 1e+06, positive = 4L, negative = 5L)
}
\arguments{
\item{file}{Path for the BAM file.}

\item{chunk}{Size of the chunk to load into memory for ATAC-seq read offset.
This argument is necessary only when \code{type="SE"}.}

\item{positive}{Positive offset: 4.}

\item{negative}{Negative offset: 5.}
}
\description{
It performs offset corrrection on single end ATAC-seq files.
}
