% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramdata.R
\docType{data}
\name{paramdata}
\alias{paramdata}
\title{Data parameters for different experiments}
\format{
A data frame with 4 rows and 6 columns
\describe{
\item{Parameter}{Name of the parameter used as input (\code{seq_type}) to the function \code{\link[=process_epigenome]{process_epigenome()}}.}
\item{Experiment}{Complete name of the experiment}
\item{Extra_bowtie2}{Extra parameters to be passed to the bowtie2 aligment.}
\item{Offset_correction}{Whether to perform (\code{TRUE}) or not (\code{FALSE}) transposase insertion offset correction.}
\item{peak_type}{Whether to call \code{narrow} or \code{broad} peaks in MACS2.}
\item{shift}{Whether to shift the coordinates of the peaks in MACS2.}
}
}
\usage{
paramdata
}
\description{
A dataset containing the different parameters used for the pipelines
analyzing different types of experiments.
}
\keyword{datasets}
