% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineNGS.R
\docType{package}
\name{pipelineNGS}
\alias{pipelineNGS}
\title{Pipeline for Next-Generation Sequencing Experiments}
\description{
This package provides an R implementation to several tools and scripts for NGS alignment and post-processing.
}
\details{
PipelineNGS is designed to facilitate the implementation of a full NGS pipeline and downstream analysis in R. It contains several wrapper functions, where arguments are passed to the command line-tools and processes are therefore run as if it were in the terminal.
}
