% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotAlignment}
\alias{plotAlignment}
\title{Plot alignment statistics}
\usage{
plotAlignment(stats, threshold = 80)
}
\arguments{
\item{stats}{Table obtained from getStats.}

\item{threshold}{Threshold for drawing the line (currently 80 percent). Indicates that all samples should have > 80 percent uniquely mapped reads.}
}
\value{
Ggplot with the alignment statistics per sample.
}
\description{
Plot alignment statistics
}
