% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCUTTAG.R
\name{processCUTTAG}
\alias{processCUTTAG}
\title{Automated processing of CUT&TAG samples}
\usage{
processCUTTAG(
  fastq_files,
  out_name,
  type = "PE",
  run_fastqc = TRUE,
  path_fastqc = "FastQC/",
  path_bam = "BAM/",
  path_peaks = "Peaks/",
  path_logs = "Logs/",
  index = "/biodata/indices/species/Hsapiens/ucsc.hg19",
  remove = c("chrM", "chrUn", "_random", "_hap", "_gl", "EBV"),
  blacklist = "~/data/consensusBlacklist.bed",
  chr_sizes = "",
  seacr_type = "stingent",
  seacr_top = 0.01,
  cores = 8,
  bedtools_bamtobed = "bedtools bamtobed",
  bedtools_genomecov = "bedtools genomecov",
  seacr = "SEACR_1.3.sh"
)
}
\arguments{
\item{fastq_files}{List containing the different FastQ file pairs (one list
element for each sample).}

\item{out_name}{Sample name to use for each of the samples.}

\item{type}{Character indicating if reads are paired end (PE) or single end
(SE).}

\item{run_fastqc}{Logical indicating whether to run FastQC or not.}

\item{path_fastqc}{Path for the FastQC results.}

\item{path_bam}{Path for the aligment and filtering results.}

\item{path_peaks}{Path for the peak files.}

\item{path_logs}{Path for the logs.}

\item{index}{Path to the index needed by Bowtie2.}

\item{remove}{List of chromosomes to remove from final BAM file.}

\item{blacklist}{List of blaclisted regions to remove from BAM file.}

\item{chr_sizes}{Chromosome sizes to use for bedgraph conversion.}

\item{seacr_type}{Type of peaks to call with SEACR, either "stringent"
(default) or "relaxed."}

\item{seacr_top}{A numeric threshold n between 0 and 1 returns the top n
fraction of peaks based on total signal within peaks. Default: 0.01.}

\item{cores}{Number of threads to use for the analysis.}

\item{bedtools_bamtobed}{Path or alias of the bedtools bamtobed utility.}

\item{bedtools_genomecov}{Path or alias of the bedtools genomecov utility.}

\item{seacr}{Path or alias of the SEACR utilty.}
}
\description{
Automated processing of CUT&TAG samples
}
