% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{tableFromIdxstats}
\alias{tableFromIdxstats}
\title{Table form idxstats}
\usage{
tableFromIdxstats(bam)
}
\arguments{
\item{bam}{Character string with the name and path of a single BAM file. Index should
be in the same folder.}
}
\value{
A data.frame with the output of samtools idxstats.
}
\description{
Given a bam file, returns samtools idxstats output as a data.frame.
}
\examples{
\dontrun{
df <- tableFromIdxstats("path/to/file.bam")
}
}
