% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauScore.R
\name{tauScore}
\alias{tauScore}
\title{Calculate tissue-specificity score}
\usage{
tauScore(counts)
}
\arguments{
\item{counts}{Integer vector of counts of a specific gene/region in different tissue samples.}
}
\value{
Tau score for tissue specificity of the specified gene/region in the specified tissues.
}
\description{
Given a list of counts for a gene/open chromatin region in different tissues, it 
returns a t score that goes from 0 (housekeeping) to 1 (tissue-specific).
}
\examples{

}
