% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeak.R
\name{callPeak}
\alias{callPeak}
\title{Peak calling with MACS2}
\usage{
callPeak(file, out_dir, path_logs, type = "narrow", shift = FALSE)
}
\arguments{
\item{file}{Character string with the filename for the BAM file.}

\item{out_dir}{Character string with the path for the peak calling result to be saved.}

\item{path_logs}{Character string with the path for the logs to be stored.}

\item{type}{Character string indicating if the resulting peaks should be "narrow" (default) or "broad".}

\item{shift}{Logical indicating wether the reads should be shifted -100bp and extended to 200bp or not (default).}
}
\value{
Bed files in path_pc with the location of the enriched regions.
}
\description{
Calls the peaks present in your sample bam files and returns bed files with the location of
the enriched regions.
}
\examples{
\dontrun{
callPeak()
}
}
