% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{Quality Control with FastQC}
\usage{
fastqc(files, path_fastqc = "fastqc/", cores = 6)
}
\arguments{
\item{files}{String or character vector with the path and filename of the
FastQ files.}

\item{path_fastqc}{Character indicating the output directory for the FastQC reports.}

\item{cores}{Number of threads to use for the analysis.}
}
\value{
Creates a folder \code{path_fastqc} containing the output reports
generated by FastQC. See the \href{https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/}{FastQC documentation}
for details.
}
\description{
Runs \href{https://www.bioinformatics.babraham.ac.uk/projects/fastqc/}{FastQC} for
quality control analysis of FastQ files.
}
\examples{
\dontrun{
fastqc(files="path/to/fastq/file.fastq.gz",
       path_fastqc="fastqc/", cores=6)
}
}
