% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtOutBAM.R
\name{filtOutBAM}
\alias{filtOutBAM}
\title{Filter out reads from BAM file}
\usage{
filtOutBAM(
  file,
  path_logs,
  type = "SE",
  remove = c("chrM", "chrUn", "_random", "_hap", "_gl", "EBV"),
  blacklist = "~/data/consensusBlacklist.bed",
  cores = 6
)
}
\arguments{
\item{file}{Filenme and path for the BAM file to be filter out.}

\item{path_logs}{Path for the output logs.}

\item{type}{Either "SE" (single end) or "PE" (paired end).}

\item{remove}{List of regular expressions to filter out from chromosome names.}

\item{blacklist}{Path for ENCODE blacklist in bed format.}

\item{cores}{Number of cores to use for the analysis.}
}
\value{
File without the ".raw" preffix containing the final reads, along with its index (.bai).
Also returns invisibly the name of the output file.
}
\description{
Function that filters out the following reads: 1) aligned no non-cannonical chromosomes, 2) not aligned, 3) aligned to ENCODE black-listed regions and 3) duplicated reads.
}
\examples{
\dontrun{
file <- filtOutBAM(file=file,
                   path_logs=path_logs,
                   cores=8)
}
}
