% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{Get alignment and processing stats from BAM files}
\usage{
getStats(raw_bam, path_logs, path_peaks, peak_type = "narrowPeak")
}
\arguments{
\item{raw_bam}{Character vector or string of BAM files generated by alignmentBowtie2() with the suffix ".raw".}

\item{path_logs}{Path where the log files from samtools mrkdup are stored.}

\item{peak_type}{Type of the peak files, either narrowPeak or broadPeak}

\item{path_peks}{Path where the peaks are saved.}
}
\value{
Data.frame with some alignment and processing statistics for all the input samples.
}
\description{
Given a single or multiple BAM files outputed by alignmentBowtie2() and filtOutBAM()
returns a data.frame with a summary of alignment and post-processing statistics.
}
\examples{
\dontrun{
stats <- getStats(raw_bam=c("path/to/file.bam", "path/to/file2.bam"), path_logs="path/logs")
}
}
