% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCUTTAG.R
\name{peakCallingSEACR}
\alias{peakCallingSEACR}
\title{CUT&TAG peak calling with SEACR}
\usage{
peakCallingSEACR(
  bam_file,
  path_peaks,
  gen_sizes,
  cores = 5,
  seacr_type = "stingent",
  seacr_top = 0.01,
  bedtools_bamtobed = "bedtools bamtobed",
  bedtools_genomecov = "bedtools genomecov",
  seacr = "SEACR_1.3.sh"
)
}
\arguments{
\item{path_peaks}{Path for the peak files.}

\item{gen_sizes}{Chromosome sizes to use for bedgraph conversion.}

\item{cores}{Number of threads to use for the analysis.}

\item{seacr_type}{Type of peaks to call with SEACR, either "stringent"
(default) or "relaxed."}

\item{seacr_top}{A numeric threshold n between 0 and 1 returns the top n
fraction of peaks based on total signal within peaks. Default: 0.01.}

\item{bedtools_bamtobed}{Path or alias of the bedtools bamtobed utility.}

\item{bedtools_genomecov}{Path or alias of the bedtools genomecov utility.}

\item{seacr}{Path or alias of the SEACR utilty.}
}
\description{
CUT&TAG peak calling with SEACR
}
