% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_epigenome}
\alias{process_epigenome}
\title{Automated processing of ChIP-seq and ATAC-seq samples}
\usage{
process_epigenome(
  fastq_files,
  seq_type = "ATAC",
  run_fastqc = TRUE,
  path_fastqc = "FastQC/",
  path_bam = "BAM/",
  path_peaks = "Peaks/",
  path_logs = "Logs/",
  out_name = NULL,
  suffix_fastq = NULL,
  type = "SE",
  index = "/biodata/indices/species/Hsapiens/ucsc.hg19",
  remove = c("chrM", "chrUn", "_random", "_hap", "_gl"),
  blacklist = "~/data/consensusBlacklist.bed",
  gen_sizes = "~/data/hg19.len",
  peak_type = NULL,
  shift = NULL,
  cores = 8
)
}
\arguments{
\item{fastq_files}{Path and filenames for the fastq files to be analyzed.}

\item{seq_type}{Type of sequencing, either ATAC (default) or CHIP.}

\item{path_fastqc}{Path for the FastQC output.}

\item{path_bam}{Path for the output BAM files.}
}
\description{
This function performs all necessary steps in the ChIP-seq processing pipeline.
}
\examples{
\dontrun{
process_epigenome(fastq_files=c("path/to/file.fastq.gz", "path/to/file2.fastq.gz"),
                  seq_type="ATAC",
                  out_name=c("sample1", "sample2"),
                  type="SE",
                  cores=8)
}

}
